Move ArgumentSanitizer into util package
Resolves #2802
